+++ src/test/java/racingcar/model/CarTest.java
@@ -1,7 +1,6 @@
-package racingcar;
+package racingcar.model;
 
 import org.junit.jupiter.api.Test;
-import racingcar.model.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
